Create using Java Forms
The BMI calculator has been developed and programmed to calculate both Metric and Imperial measures and the user has to input the following.
Male/Female: The user need to select male or female, these have been created as radio buttons
Age: Once age is selected the radio buttons for imperial and metric are enabled, a calculation cannot be performed without this text box being completed. If the age is over or equal to 110 text stating incorrect age will appear.
Imperial/Metric: The user needs to select the unit that they would like to enter, once a unit has been chosen the option not chosen will be disabled also the calculate button for the unit that they have chosen will now be enabled.
Calculate: Once all of the measurements have been completed, the text at the bottom of the screen will show BMI total and category.
Clear – The user can clear all of the entries and options that they have chosen and the BMI calculator will reset to its origin state.